Moved TelemetryConfig to Core
