+++ src/Altinn.AccessManagement.Core/Altinn.AccessManagement.Core.csproj
@@ -20,6 +20,7 @@
     <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="8.0.0" />
     <PackageReference Include="Npgsql" Version="8.0.2" />
+    <PackageReference Include="OpenTelemetry" Version="1.7.0" />
   </ItemGroup>
 
   <ItemGroup>
