--- src/Altinn.AccessManagement.Persistence/Configuration/TelemetryConfig.cs
@@ -1,23 +1,0 @@
-﻿using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
-using System.Diagnostics.Metrics;
-
-namespace Altinn.AccessManagement.Persistence.Configuration
-{
-    /// <summary>
-    /// Config to be used for Telemetry in Altinn.AccessManagement.Persistence
-    /// </summary>
-    [ExcludeFromCodeCoverage]
-    public static class TelemetryConfig
-    {
-        /// <summary>
-        /// Used as source for the current project
-        /// </summary>
-        public static readonly ActivitySource _activitySource = new("Altinn.AccessManagement.Persistence");
-
-        /// <summary>
-        /// Used as source for the current project
-        /// </summary>
-        public static readonly Meter _meter = new("Altinn.AccessManagement.Persistence");
-    }
-}
