+++ src/Altinn.AccessManagement.Persistence/DelegationMetadataRepository.cs
@@ -1,10 +1,12 @@
 using System.Data;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
+using Altinn.AccessManagement.Core.Constants;
 using Altinn.AccessManagement.Core.Enums;
 using Altinn.AccessManagement.Core.Models;
 using Altinn.AccessManagement.Core.Models.ResourceRegistry;
 using Altinn.AccessManagement.Core.Repositories.Interfaces;
+using Altinn.AccessManagement.Core.Telemetry;
 using Altinn.AccessManagement.Persistence.Extensions;
 using Npgsql;
 using NpgsqlTypes;
@@ -70,7 +72,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
     /// <inheritdoc/>
     public async Task<List<DelegationChange>> GetAllAppDelegationChanges(string altinnAppId, int offeredByPartyId, int? coveredByPartyId, int? coveredByUserId)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getAllAppDelegationChanges);
@@ -146,7 +148,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
     /// <inheritdoc/>
     public async Task<List<DelegationChange>> GetOfferedResourceRegistryDelegations(int offeredByPartyId, List<string> resourceRegistryIds = null, List<ResourceType> resourceTypes = null, CancellationToken cancellationToken = default)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getResourceRegistryDelegationChangesOfferedByPartyId);
@@ -175,7 +177,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
     /// <inheritdoc/>
     public async Task<List<DelegationChange>> GetReceivedResourceRegistryDelegationsForCoveredByPartys(List<int> coveredByPartyIds, List<int> offeredByPartyIds = null, List<string> resourceRegistryIds = null, List<ResourceType> resourceTypes = null)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getResourceRegistryDelegationChangesForCoveredByPartyIds);
@@ -205,7 +207,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
     /// <inheritdoc/>
     public async Task<List<DelegationChange>> GetReceivedResourceRegistryDelegationsForCoveredByUser(int coveredByUserId, List<int> offeredByPartyIds, List<string> resourceRegistryIds = null, List<ResourceType> resourceTypes = null)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getResourceRegistryDelegationChangesForCoveredByUserId);
@@ -235,7 +237,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
     /// <inheritdoc/>
     public async Task<List<DelegationChange>> GetOfferedDelegations(List<int> offeredByPartyIds, CancellationToken cancellationToken = default)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
 
         const string QUERY = /*strpsql*/@"
             WITH resources AS (
@@ -320,7 +322,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
     /// <inheritdoc/>
     public async Task<List<DelegationChange>> GetAllDelegationChangesForAuthorizedParties(List<int> coveredByUserIds, List<int> coveredByPartyIds, CancellationToken cancellationToken = default)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
 
         if (coveredByUserIds == null && coveredByPartyIds == null)
         {
@@ -414,7 +416,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private async Task<DelegationChange> InsertAppDelegation(DelegationChange delegationChange)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(insertAppDelegationChange);
@@ -445,7 +447,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private async Task<DelegationChange> InsertResourceRegistryDelegation(DelegationChange delegationChange)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(insertResourceRegistryDelegationChange);
@@ -478,7 +480,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private async Task<DelegationChange> GetCurrentAppDelegation(string resourceId, int offeredByPartyId, int? coveredByPartyId, int? coveredByUserId)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
 
         try
         {
@@ -506,7 +508,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private async Task<DelegationChange> GetCurrentResourceRegistryDelegation(string resourceId, int offeredByPartyId, int? coveredByPartyId, int? coveredByUserId)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getCurrentResourceRegistryDelegationChange);
@@ -534,7 +536,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
     /// <inheritdoc/>
     public async Task<List<DelegationChange>> GetResourceRegistryDelegationChanges(List<string> resourceIds, int offeredByPartyId, int coveredByPartyId, ResourceType resourceType)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getResourceRegistryDelegationChanges);
@@ -583,7 +585,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private static async ValueTask<DelegationChange> GetAppDelegationChange(NpgsqlDataReader reader)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity();
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity();
         try
         {
             return new DelegationChange
@@ -610,7 +612,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private static async ValueTask<DelegationChange> GetResourceRegistryDelegationChange(NpgsqlDataReader reader)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity();
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity();
         try
         {
             return new DelegationChange
@@ -638,7 +640,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private async Task<List<DelegationChange>> GetAllCurrentAppDelegationChangesCoveredByPartyIds(List<string> altinnAppIds = null, List<int> offeredByPartyIds = null, List<int> coveredByPartyIds = null)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getAppDelegationChangesForCoveredByPartyIds);
@@ -667,7 +669,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private async Task<List<DelegationChange>> GetAllCurrentAppDelegationChangesCoveredByUserIds(List<string> altinnAppIds = null, List<int> offeredByPartyIds = null, List<int> coveredByUserIds = null)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getAppDelegationChangesForCoveredByUserIds);
@@ -696,7 +698,7 @@ public class DelegationMetadataRepository : IDelegationMetadataRepository
 
     private async Task<List<DelegationChange>> GetAllCurrentAppDelegationChangesOfferedByPartyIdOnly(List<string> altinnAppIds = null, List<int> offeredByPartyIds = null)
     {
-        using var activity = Configuration.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
+        using var activity = Core.Telemetry.TelemetryConfig._activitySource.StartActivity(ActivityKind.Client);
         try
         {
             await using var pgcom = _conn.CreateCommand(getAppDelegationChangesOfferedByPartyIds);
