+++ src/Altinn.AccessManagement.Persistence/ResourceMetadataRepository.cs
@@ -4,6 +4,7 @@ using System.Diagnostics.CodeAnalysis;
 using Altinn.AccessManagement.Core.Models;
 using Altinn.AccessManagement.Core.Models.ResourceRegistry;
 using Altinn.AccessManagement.Core.Repositories.Interfaces;
+using Altinn.AccessManagement.Core.Telemetry;
 using Altinn.AccessManagement.Persistence.Configuration;
 using Altinn.AccessManagement.Persistence.Extensions;
 using Microsoft.Extensions.Options;
